Revert "refactor(session): split CheckpointManager into its own file (#511)"
This reverts commit d65320b1cf873edc94ed3bd8a52ed110f2ebd66c.
