+++ src/session/mod.rs
@@ -11,7 +11,6 @@
 //! - [`bootstrap`] — Bootstrap context protection during compaction
 
 pub mod bootstrap;
-pub mod checkpoint_manager;
 pub mod events;
 pub mod persistence;
 #[cfg(test)]
@@ -22,6 +21,7 @@ pub mod sweeper;
 
 // Re-export commonly used types
 pub use bootstrap::{BootstrapContext, BootstrapProtection, BootstrapRegion};
-pub use checkpoint_manager::CheckpointManager;
 pub use events::{CheckpointTrigger, ModeSwitchEvent, UserIntent};
-pub use persistence::{PersistenceError, PersistenceService, ReasoningMode, SessionCheckpoint};
+pub use persistence::{
+    CheckpointManager, PersistenceError, PersistenceService, ReasoningMode, SessionCheckpoint,
+};
