+++ src/session/persistence_tests.rs
@@ -7,11 +7,10 @@ mod tests {
     use chrono::Utc;
 
     use crate::session::persistence::{
-        PendingMessage, PersistenceService, ReasoningMode, ReasoningModeState, SessionCheckpoint,
-        SessionStatus,
+        CheckpointManager, PendingMessage, PersistenceService, ReasoningMode, ReasoningModeState,
+        SessionCheckpoint, SessionStatus,
     };
     use crate::session::storage::memory::MemoryStorage;
-    use crate::session::CheckpointManager;
 
     fn create_test_checkpoint(session_id: &str) -> SessionCheckpoint {
         let mut state = ReasoningModeState::default();
